DEV-17017 Reorganize and simplify dependencies
First pass at this, kinda messy but working and way nicer import API.

Move standard C FFI functionality to `ffi_core`. Strip everything else
out of `ffi_common`; repurpose it as a top-level crate that reexports
`ffi_core`, `ffi_consumer`, and `ffi_derive`.

Update proc macros to use reexports from `ffi_common`. Picked up a fun
trick from serde for tests -- `ffi_derive` can have a dev-dependency on
`ffi_common`, which will let the tests live in `ffi_derive` while still
using the `ffi_common` import paths.

With these changes, consumer crates only have to depend on `ffi_common`
(and they still don't have access to any of the token parsing madness in
`ffi_internals`).
