+++ Cargo.toml
@@ -2,6 +2,7 @@
 members = [
     "ffi_common",
     "ffi_consumer",
+    "ffi_core",
     "ffi_derive",
     "ffi_internals"
 ]
@@ -15,6 +16,12 @@ path = "./ffi_common"
 path = "./ffi_consumer"
 # git = "ssh://git@github.com/agrian-inc/ffi_common.git"
 # branch = "develop"
+
+# [patch."https://github.com/agrian-inc/ffi_common"]
+# [patch."agrian-registry".ffi_core]
+# path = "./ffi_core"
+# # git = "ssh://git@github.com/agrian-inc/ffi_common.git"
+# # branch = "develop"
 
 [patch."agrian-registry".ffi_derive]
 path = "./ffi_derive"
