+++ ffi_common/src/lib.rs
@@ -1,6 +1,8 @@
 //! # `ffi_common`
 //!
-//! Crate for common FFI behaviors needed by other Rust crates that provide a C interface.
+//! Crate for common Rust FFI behaviors, including error, string, and primitive handling in
+//! `ffi_internals`, generating an ffi with `ffi_derive`, and generating consumer types around that
+//! FFI with `ffi_consumer`.
 //!
 
 #![warn(
@@ -20,30 +22,6 @@
 )]
 #![forbid(missing_docs, unused_extern_crates, unused_imports)]
 
-#[macro_use]
-pub mod error;
-pub mod datetime;
-pub mod macros;
-pub mod string;
-
-pub use paste::paste;
-
-declare_value_type_ffi!(bool, u8, u16, u32, u64, i8, i16, i32, i64, f32, f64);
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_ffi_none() {
-        let null_pointer = option_u8_init(false, 0);
-        assert!(null_pointer.is_null());
-    }
-
-    #[test]
-    fn test_ffi_some() {
-        let u8_pointer = option_u8_init(true, 3);
-        assert!(!u8_pointer.is_null());
-        assert_eq!(unsafe { *Box::from_raw(u8_pointer as *mut u8) }, 3);
-    }
-}
+pub use ffi_derive;
+pub use ffi_consumer;
+pub use ffi_core;
