+++ ffi_consumer/Cargo.toml
@@ -8,7 +8,6 @@ repository = "https://github.com/agrian-inc/ffi_common"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-ffi_common = { version = "0.3", registry = "agrian-registry" }
 ffi_internals = { version = "0.3", registry = "agrian-registry" }
 heck = "0.3"
 syn = { version = "1.0", features = ["full"] }
