+++ ffi_core/src/error.rs
