+++ ffi_internals/src/struct_internals/struct_ffi.rs
@@ -120,7 +120,7 @@ impl From<StructFFI> for TokenStream {
         quote!(
             #[allow(box_pointers, missing_docs)]
             pub mod #module_name {
-                use ffi_common::{*, datetime::*, ffi_string, paste, string::FFIArrayString};
+                use ffi_common::ffi_core::{*, paste, datetime::*, string::FFIArrayString};
                 use std::os::raw::c_char;
                 use std::{ffi::{CStr, CString}, mem::ManuallyDrop, ptr};
                 #extra_imports
